Mon, Nov 28, 2022, 12:58 PM +00:00

+++ filter-less/helpers.c
@@ -108,9 +108,9 @@ RGBTRIPLE blurredPixel(int height, int width, RGBTRIPLE copy[height][width], int
     int startY = y;
     int endY = y;
     int counter = 0;
-    int averageBlue;
-    int averageGreen;
-    int averageRed;
+    double averageBlue;
+    double averageGreen;
+    double averageRed;
     // if (x!=0)
     // {
     //     startX = x-1;
